+++ R/download_netlogo.R
@@ -18,11 +18,10 @@
 #' * "5.3.1" = NetLogo Version 5.3.1
 #'
 #'
-#'
 #' @examples
 #' \donttest{
 #' dlpath <- tempdir()  # adjust path to your needs
-#' download_netlogo(dlpath, "6.0.3")
+#' try(download_netlogo(dlpath, "6.0.3"))
 #' }
 #'
 #' @aliases download_netlogo