+++ R/util_runnl.R
@@ -342,14 +342,16 @@ util_gather_results <- function(nl, outfile, seed, siminputrow) {
 ## Clean patch metrics
 .util_clean_metrics_patches <- function(NLtable, nl) {
 
-  patches_string <- NLtable[, grepl(c("metrics.patches"), names(NLtable))]
+  patches_string <- NLtable[, grepl(c("metrics.patches"), names(NLtable))]  %>%
+    dplyr::mutate_all(function(x) gsub('[\"]', '',x))
+
   patches_string <- stringr::str_split(patches_string$metrics.patches, " ")
   patches_string <- purrr::map(patches_string, function(x) {
     patches_owns <- tibble::as_tibble(x = x)
     patches_owns <- tidyr::separate(patches_owns, value,
                                     getexp(nl, "metrics.patches"), sep=",")
     patches_owns <- dplyr::mutate_all(patches_owns, function(x) {
-      suppressWarnings(if(!is.na(x)) {ifelse(is.na(as.numeric(as.character(x))),
+      suppressWarnings(if(!all(is.na(x))) {ifelse(is.na(as.numeric(as.character(x))),
                               as.character(x),
                               as.numeric(as.character(x)))})
     })
@@ -363,7 +365,9 @@ util_gather_results <- function(nl, outfile, seed, siminputrow) {
 
 .util_clean_metrics_turtles <- function(NLtable, nl, col.name, metrics) {
 
-  turtles_string <- NLtable[, grepl(col.name, names(NLtable))]
+  turtles_string <- NLtable[, grepl(col.name, names(NLtable))]  %>%
+    dplyr::mutate_all(function(x) gsub('[\"]', '',x))
+
   turtles_string <- stringr::str_split(dplyr::pull(turtles_string, col.name), " ")
   turtles_string <- purrr::map(turtles_string, function(x) {
     turtles_owns <- tibble::as_tibble(x = x)
@@ -372,7 +376,7 @@ util_gather_results <- function(nl, outfile, seed, siminputrow) {
                                     metrics,
                                     sep=",")
     turtles_owns <- dplyr::mutate_all(turtles_owns, function(x) {
-      suppressWarnings(if(!is.na(x)) {ifelse(is.na(as.numeric(as.character(x))),
+      suppressWarnings(if(!all(is.na(x))) {ifelse(is.na(as.numeric(as.character(x))),
                               as.character(x),
                               as.numeric(as.character(x)))})
     })
@@ -385,7 +389,9 @@ util_gather_results <- function(nl, outfile, seed, siminputrow) {
 # nocov start
 .util_clean_metrics_links <- function(NLtable, nl, col.name, metrics) {
 
-  links_string <- NLtable[, grepl(col.name, names(NLtable))]
+  links_string <- NLtable[, grepl(col.name, names(NLtable))]  %>%
+    dplyr::mutate_all(function(x) gsub('[\"]', '',x))
+
   links_string <- stringr::str_split(dplyr::pull(links_string, col.name), " ")
   links_string <- purrr::map(links_string, function(x) {
     links_owns <- tibble::as_tibble(x = x)
@@ -394,7 +400,7 @@ util_gather_results <- function(nl, outfile, seed, siminputrow) {
                                   metrics,
                                   sep=",")
     links_owns <- dplyr::mutate_all(links_owns, function(x) {
-      suppressWarnings(if(!is.na(x)) {ifelse(is.na(as.numeric(as.character(x))),
+      suppressWarnings(if(!all(is.na(x))) {ifelse(is.na(as.numeric(as.character(x))),
                               as.character(x),
                               as.numeric(as.character(x)))})
 
@@ -566,6 +572,9 @@ util_read_write_batch <- function(nl) {
       system(paste0("sed -i -r 's!^JVM_OPTS=.*!",
                     jvmoptsline, "!'", " \"",
                     batchpath_temp, "\""))
+
+      ## Make sh executable on linux:
+      system(paste0("chmod +x ", batchpath_temp), wait = TRUE)
     }
   }